refactor: phase 3 - move route composition into view modules
