+++ app/graphql/gql.ts
@@ -17,45 +17,45 @@ type Documents = {
     "\n  query EventContentSchedule($eventUid: String!) {\n    eventContent(uid: $eventUid) {\n      schedules {\n        region\n        runType\n        startAt\n        endAt\n      }\n    }\n  }\n": typeof types.EventContentScheduleDocument,
     "\n  query EventContentsList {\n    eventContents {\n      uid\n      name\n      schedules {\n        region\n        runType\n        startAt\n        endAt\n      }\n    }\n  }\n": typeof types.EventContentsListDocument,
     "\n  query EventContentShopContent($eventUid: String!, $runType: RunTypeEnum!) {\n    eventContent(uid: $eventUid) {\n      stages(runType: $runType) {\n        uid stageNumber stageIndex stageType enterCostAmount\n        rewards {\n          amount probability tag\n          resource { __typename uid name rarity ... on Item { category } }\n        }\n      }\n      shopResources(runType: $runType) {\n        uid resourceAmount shopAmount\n        resource { type uid name rarity }\n        paymentResource { type uid name }\n        purchaseTiers {\n          tierIndex\n          startQuantity\n          quantity\n          unitPrice\n          paymentResource { type uid name }\n        }\n      }\n      bonuses(runType: $runType) {\n        percentage\n        resource { uid name }\n        student { uid name role }\n      }\n      minigameConfigs(runType: $runType) {\n        minigameType\n        payment { quantity resource { type uid name } }\n        payments { quantity resource { type uid name } }\n        rewardGroups {\n          condition { type value values divisor remainders }\n          payments {\n            quantityMin\n            quantityExpected\n            quantityMax\n            quantityVariable\n            resource { type uid name }\n          }\n          rewards { quantity resource { type uid name rarity } }\n        }\n      }\n    }\n  }\n": typeof types.EventContentShopContentDocument,
-    "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n": typeof types.MainStoriesDocument,
-    "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n": typeof types.RaidScheduleDetailDocument,
-    "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n": typeof types.AllRaidSchedulesDocument,
-    "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": typeof types.AllStudentsFavoriteItemsDocument,
-    "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n": typeof types.AllStudentsDocument,
-    "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n": typeof types.StudentSkillItemsDocument,
     "\n  query GrowthSkillCosts($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      ex2: skillItems(skillType: ex, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex3: skillItems(skillType: ex, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex4: skillItems(skillType: ex, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex5: skillItems(skillType: ex, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal2: skillItems(skillType: normal, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal3: skillItems(skillType: normal, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal4: skillItems(skillType: normal, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal5: skillItems(skillType: normal, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal6: skillItems(skillType: normal, skillLevel: 6) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal7: skillItems(skillType: normal, skillLevel: 7) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal8: skillItems(skillType: normal, skillLevel: 8) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal9: skillItems(skillType: normal, skillLevel: 9) { amount item { uid rarity ... on Item { category subCategory } } }\n    }\n  }\n": typeof types.GrowthSkillCostsDocument,
     "\n  query GrowthResourceItems($uids: [String!]) {\n    items(uids: $uids) {\n      uid name rarity type\n      ... on Item { category subCategory }\n    }\n  }\n": typeof types.GrowthResourceItemsDocument,
     "\n  query GrowthResourceEquipments($uids: [String!]) {\n    equipments(uids: $uids) {\n      uid name rarity type category\n    }\n  }\n": typeof types.GrowthResourceEquipmentsDocument,
     "\n  query GrowthStudentGears($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      gear {\n        name\n        growthItems {\n          gearTier\n          amount\n          item {\n            uid name rarity type\n            ... on Item { category subCategory }\n          }\n        }\n      }\n    }\n  }\n": typeof types.GrowthStudentGearsDocument,
     "\n  query UserResourceInventoryCatalog {\n    items {\n      uid name rarity type\n      ... on Item { category subCategory }\n    }\n    equipments {\n      uid name rarity type category\n    }\n  }\n": typeof types.UserResourceInventoryCatalogDocument,
+    "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n": typeof types.MainStoriesDocument,
+    "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n": typeof types.RaidScheduleDetailDocument,
+    "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n": typeof types.AllRaidSchedulesDocument,
     "\n  query RecruitmentGroupsList($endAfter: ISO8601DateTime, $uids: [String!]) {\n    recruitmentGroups(endAfter: $endAfter, uids: $uids) {\n      uid contentType contentUid startAt endAt recruitmentType\n      recruitments {\n        recruitmentType pickup rerun since until studentName\n        student { uid attackType defenseType role name schaleDbId initialTier releaseAt archiveAt }\n      }\n    }\n  }\n": typeof types.RecruitmentGroupsListDocument,
     "\n  query RecruitmentPoolStudents {\n    students {\n      uid name initialTier attackType defenseType role releaseAt archiveAt\n      recruitments {\n        recruitmentType\n      }\n    }\n  }\n": typeof types.RecruitmentPoolStudentsDocument,
-    "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n": typeof types.FarmingStagesDocument,
+    "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": typeof types.AllStudentsFavoriteItemsDocument,
     "\n  query StudentFavoriteItem($uid: String!) {\n    student(uid: $uid) {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": typeof types.StudentFavoriteItemDocument,
-    "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n": typeof types.StudentGradeDetailDocument,
+    "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n": typeof types.FarmingStagesDocument,
+    "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n": typeof types.AllStudentsDocument,
     "\n  query StudentDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId releaseAt\n      recruitments {\n        since rerun\n        recruitmentGroup { uid startAt endAt }\n      }\n    }\n  }\n": typeof types.StudentDetailDocument,
+    "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n": typeof types.StudentGradeDetailDocument,
+    "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n": typeof types.StudentSkillItemsDocument,
 };
 const documents: Documents = {
     "\n  query EventContentSchedule($eventUid: String!) {\n    eventContent(uid: $eventUid) {\n      schedules {\n        region\n        runType\n        startAt\n        endAt\n      }\n    }\n  }\n": types.EventContentScheduleDocument,
     "\n  query EventContentsList {\n    eventContents {\n      uid\n      name\n      schedules {\n        region\n        runType\n        startAt\n        endAt\n      }\n    }\n  }\n": types.EventContentsListDocument,
     "\n  query EventContentShopContent($eventUid: String!, $runType: RunTypeEnum!) {\n    eventContent(uid: $eventUid) {\n      stages(runType: $runType) {\n        uid stageNumber stageIndex stageType enterCostAmount\n        rewards {\n          amount probability tag\n          resource { __typename uid name rarity ... on Item { category } }\n        }\n      }\n      shopResources(runType: $runType) {\n        uid resourceAmount shopAmount\n        resource { type uid name rarity }\n        paymentResource { type uid name }\n        purchaseTiers {\n          tierIndex\n          startQuantity\n          quantity\n          unitPrice\n          paymentResource { type uid name }\n        }\n      }\n      bonuses(runType: $runType) {\n        percentage\n        resource { uid name }\n        student { uid name role }\n      }\n      minigameConfigs(runType: $runType) {\n        minigameType\n        payment { quantity resource { type uid name } }\n        payments { quantity resource { type uid name } }\n        rewardGroups {\n          condition { type value values divisor remainders }\n          payments {\n            quantityMin\n            quantityExpected\n            quantityMax\n            quantityVariable\n            resource { type uid name }\n          }\n          rewards { quantity resource { type uid name rarity } }\n        }\n      }\n    }\n  }\n": types.EventContentShopContentDocument,
-    "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n": types.MainStoriesDocument,
-    "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n": types.RaidScheduleDetailDocument,
-    "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n": types.AllRaidSchedulesDocument,
-    "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": types.AllStudentsFavoriteItemsDocument,
-    "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n": types.AllStudentsDocument,
-    "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n": types.StudentSkillItemsDocument,
     "\n  query GrowthSkillCosts($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      ex2: skillItems(skillType: ex, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex3: skillItems(skillType: ex, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex4: skillItems(skillType: ex, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex5: skillItems(skillType: ex, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal2: skillItems(skillType: normal, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal3: skillItems(skillType: normal, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal4: skillItems(skillType: normal, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal5: skillItems(skillType: normal, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal6: skillItems(skillType: normal, skillLevel: 6) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal7: skillItems(skillType: normal, skillLevel: 7) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal8: skillItems(skillType: normal, skillLevel: 8) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal9: skillItems(skillType: normal, skillLevel: 9) { amount item { uid rarity ... on Item { category subCategory } } }\n    }\n  }\n": types.GrowthSkillCostsDocument,
     "\n  query GrowthResourceItems($uids: [String!]) {\n    items(uids: $uids) {\n      uid name rarity type\n      ... on Item { category subCategory }\n    }\n  }\n": types.GrowthResourceItemsDocument,
     "\n  query GrowthResourceEquipments($uids: [String!]) {\n    equipments(uids: $uids) {\n      uid name rarity type category\n    }\n  }\n": types.GrowthResourceEquipmentsDocument,
     "\n  query GrowthStudentGears($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      gear {\n        name\n        growthItems {\n          gearTier\n          amount\n          item {\n            uid name rarity type\n            ... on Item { category subCategory }\n          }\n        }\n      }\n    }\n  }\n": types.GrowthStudentGearsDocument,
     "\n  query UserResourceInventoryCatalog {\n    items {\n      uid name rarity type\n      ... on Item { category subCategory }\n    }\n    equipments {\n      uid name rarity type category\n    }\n  }\n": types.UserResourceInventoryCatalogDocument,
+    "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n": types.MainStoriesDocument,
+    "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n": types.RaidScheduleDetailDocument,
+    "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n": types.AllRaidSchedulesDocument,
     "\n  query RecruitmentGroupsList($endAfter: ISO8601DateTime, $uids: [String!]) {\n    recruitmentGroups(endAfter: $endAfter, uids: $uids) {\n      uid contentType contentUid startAt endAt recruitmentType\n      recruitments {\n        recruitmentType pickup rerun since until studentName\n        student { uid attackType defenseType role name schaleDbId initialTier releaseAt archiveAt }\n      }\n    }\n  }\n": types.RecruitmentGroupsListDocument,
     "\n  query RecruitmentPoolStudents {\n    students {\n      uid name initialTier attackType defenseType role releaseAt archiveAt\n      recruitments {\n        recruitmentType\n      }\n    }\n  }\n": types.RecruitmentPoolStudentsDocument,
-    "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n": types.FarmingStagesDocument,
+    "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": types.AllStudentsFavoriteItemsDocument,
     "\n  query StudentFavoriteItem($uid: String!) {\n    student(uid: $uid) {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n": types.StudentFavoriteItemDocument,
-    "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n": types.StudentGradeDetailDocument,
+    "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n": types.FarmingStagesDocument,
+    "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n": types.AllStudentsDocument,
     "\n  query StudentDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId releaseAt\n      recruitments {\n        since rerun\n        recruitmentGroup { uid startAt endAt }\n      }\n    }\n  }\n": types.StudentDetailDocument,
+    "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n": types.StudentGradeDetailDocument,
+    "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n": types.StudentSkillItemsDocument,
 };
 
 /**
@@ -87,30 +87,6 @@ export function graphql(source: "\n  query EventContentShopContent($eventUid: St
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n"): (typeof documents)["\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n"): (typeof documents)["\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n"): (typeof documents)["\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n"): (typeof documents)["\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n"): (typeof documents)["\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
 export function graphql(source: "\n  query GrowthSkillCosts($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      ex2: skillItems(skillType: ex, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex3: skillItems(skillType: ex, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex4: skillItems(skillType: ex, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex5: skillItems(skillType: ex, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal2: skillItems(skillType: normal, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal3: skillItems(skillType: normal, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal4: skillItems(skillType: normal, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal5: skillItems(skillType: normal, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal6: skillItems(skillType: normal, skillLevel: 6) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal7: skillItems(skillType: normal, skillLevel: 7) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal8: skillItems(skillType: normal, skillLevel: 8) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal9: skillItems(skillType: normal, skillLevel: 9) { amount item { uid rarity ... on Item { category subCategory } } }\n    }\n  }\n"): (typeof documents)["\n  query GrowthSkillCosts($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      ex2: skillItems(skillType: ex, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex3: skillItems(skillType: ex, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex4: skillItems(skillType: ex, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      ex5: skillItems(skillType: ex, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal2: skillItems(skillType: normal, skillLevel: 2) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal3: skillItems(skillType: normal, skillLevel: 3) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal4: skillItems(skillType: normal, skillLevel: 4) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal5: skillItems(skillType: normal, skillLevel: 5) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal6: skillItems(skillType: normal, skillLevel: 6) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal7: skillItems(skillType: normal, skillLevel: 7) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal8: skillItems(skillType: normal, skillLevel: 8) { amount item { uid rarity ... on Item { category subCategory } } }\n      normal9: skillItems(skillType: normal, skillLevel: 9) { amount item { uid rarity ... on Item { category subCategory } } }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
@@ -131,6 +107,18 @@ export function graphql(source: "\n  query UserResourceInventoryCatalog {\n    i
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
+export function graphql(source: "\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query MainStories {\n    mainStories {\n      uid name label season sortOrder\n      chapters {\n        uid name chapterNumber\n        parts {\n          uid name episodeStart episodeEnd sortOrder\n          schedules { region releasedAt confirmed }\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n"): (typeof documents)["\n  query RaidScheduleDetail($uid: String!) {\n    raidSchedule(uid: $uid) {\n      uid raidType seasonIndex region terrain startAt endAt attackType\n      raidBoss { uid name }\n      defenseTypeSets { difficulty defenseTypes }\n      jpSchedule { uid seasonIndex startAt }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n"): (typeof documents)["\n  query AllRaidSchedules($region: String!, $endAfter: ISO8601DateTime) {\n    raidSchedules(region: $region, endAfter: $endAfter) {\n      nodes {\n        uid raidType seasonIndex region terrain startAt endAt attackType\n        raidBoss { uid name }\n        defenseTypeSets { difficulty defenseTypes }\n        jpSchedule { uid seasonIndex startAt }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
 export function graphql(source: "\n  query RecruitmentGroupsList($endAfter: ISO8601DateTime, $uids: [String!]) {\n    recruitmentGroups(endAfter: $endAfter, uids: $uids) {\n      uid contentType contentUid startAt endAt recruitmentType\n      recruitments {\n        recruitmentType pickup rerun since until studentName\n        student { uid attackType defenseType role name schaleDbId initialTier releaseAt archiveAt }\n      }\n    }\n  }\n"): (typeof documents)["\n  query RecruitmentGroupsList($endAfter: ISO8601DateTime, $uids: [String!]) {\n    recruitmentGroups(endAfter: $endAfter, uids: $uids) {\n      uid contentType contentUid startAt endAt recruitmentType\n      recruitments {\n        recruitmentType pickup rerun since until studentName\n        student { uid attackType defenseType role name schaleDbId initialTier releaseAt archiveAt }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
@@ -139,7 +127,7 @@ export function graphql(source: "\n  query RecruitmentPoolStudents {\n    studen
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n"): (typeof documents)["\n  query AllStudentsFavoriteItems {\n    students {\n      uid name\n      favoriteItems {\n        favorited favoriteLevel exp\n        item { uid name rarity }\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -147,11 +135,23 @@ export function graphql(source: "\n  query StudentFavoriteItem($uid: String!) {\
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n"): (typeof documents)["\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n"];
+export function graphql(source: "\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query FarmingStages($category: String) {\n    stages(category: $category) {\n      uid name stageNumber area difficulty terrain\n      entryCosts { amount resource { uid name type } }\n      rewards(region: gl) {\n        rewardType rewardTag probability\n        resource { __typename uid name type rarity ... on Equipment { category } }\n        gachaGroup {\n          items(region: gl) {\n            chance\n            resource { __typename uid name type rarity ... on Equipment { category } }\n          }\n        }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n"): (typeof documents)["\n  query AllStudents {\n    students {\n      uid name familyName altNames school initialTier order attackType defenseType position tacticRole birthday role equipments released\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query StudentDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId releaseAt\n      recruitments {\n        since rerun\n        recruitmentGroup { uid startAt endAt }\n      }\n    }\n  }\n"): (typeof documents)["\n  query StudentDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId releaseAt\n      recruitments {\n        since rerun\n        recruitmentGroup { uid startAt endAt }\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n"): (typeof documents)["\n  query StudentGradeDetail($uid: String!) {\n    student(uid: $uid) {\n      name familyName uid attackType defenseType role school schaleDbId\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n"): (typeof documents)["\n  query StudentSkillItems($uids: [String!]) {\n    students(uids: $uids) {\n      uid\n      schaleDbId\n      skillItems(skillType: ex, skillLevel: 5) {\n        item { uid subCategory rarity }\n      }\n    }\n  }\n"];
 
 export function graphql(source: string) {
   return (documents as any)[source] ?? {};
