+++ app/routes/[__manage].tsx
@@ -3,8 +3,7 @@ import type { ActionFunctionArgs, LoaderFunctionArgs, MetaFunction } from "react
 import { Form, data, redirect, useActionData, useNavigation } from "react-router";
 import { getActiveSensei } from "~/auth/authenticator.server";
 import { Button, Callout, Title } from "~/components/primitives";
-import { getFutureContents, getIndexContents, getNavigationBarContentsRaw } from "~/models/content";
-import { getEventList, syncEventContentsList, warmActiveUpcomingEventContent } from "~/models/event-content";
+import { syncEventContentsList } from "~/models/event-content";
 import { getMainStories } from "~/models/main-story";
 import { warmRaidCache } from "~/models/raid";
 import { getAllStudentsFavoriteItems } from "~/models/resource";
@@ -16,6 +15,10 @@ import { getStudentGearData } from "~/models/growth-resource";
 import { getItemCatalogResources } from "~/models/item-catalog";
 import { warmRecruitmentCache } from "~/models/recruitment";
 import { getCampaignFarmingStages } from "~/models/stage";
+import { getEventList, warmActiveUpcomingEventContent } from "~/views/events";
+import { getFutureContents } from "~/views/futures";
+import { getIndexContents } from "~/views/home";
+import { getNavigationBarContentsRaw } from "~/views/navigation";
 
 type RefreshTaskName =
   | "syncYoutubeCommunityPosts"
