+++ app/routes/api.recruitment-results.tsx
@@ -2,7 +2,7 @@ import type { ActionFunctionArgs } from "react-router";
 import { data } from "react-router";
 import { getActiveSensei } from "~/auth/authenticator.server";
 import { nowUtcIso } from "~/lib/date-time";
-import { getFutureContents } from "~/models/content";
+import { getFutureContents } from "~/views/futures";
 import { getUserFavoritedStudents } from "~/models/favorite-students";
 import { canCompleteRecruitmentStudent } from "~/models/recruitment-result-completion";
 import {
